+++ src/Rocketeer/Traits/BashModules/Scm.php
@@ -43,8 +43,10 @@ class Scm extends Binaries
 		}
 
 		// Deploy submodules
-		$this->command->info('Initializing submodules if any');
-		$this->runForCurrentRelease($this->scm->submodules());
+		if ($this->rocketeer->getOption('scm.submodules')) {
+			$this->command->info('Initializing submodules if any');
+			$this->runForCurrentRelease($this->scm->submodules());
+		}
 
 		return $success;
 	}