+++ src/config/remote.php
@@ -14,7 +14,7 @@
 	'root_directory'   => '/home/www/',
 
 	// The process that will be executed by Composer
-	'composer' => function($task) {
+	'composer' => function ($task) {
 		return array(
 			$task->composer('self-update'),
 			$task->composer('install --no-interaction --no-dev --prefer-dist'),