+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added various SSH task-running helpers such as `Rocketeer::task(taskname, task)`
 - Added an option to disable recursive git clone (submodules)
+- Added various Composer-related options
+- Releases are now sorted by date when printed out in `rollback` and `current`
 
 ### 1.1.2